tests/requirements/base.txt
coverage
django-app-helper
tox
coverage
isort
flake8
pyflakes>=2.1
wheel
black
pre-commit
djangocms-text
django-filer
html5lib
-e .
pytest
pytest-django
